Release DHCP cleanly when using dhcpcd

+++ platform/kobo/release-ip.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
 # Release IP and shutdown udhcpc.
-killall udhcpc default.script dhcpcd 2>/dev/null
-ifconfig "${INTERFACE}" 0.0.0.0
+if [ -x "/sbin/dhcpcd" ]; then
+    env -u LD_LIBRARY_PATH dhcpcd -d -k "${INTERFACE}"
+else
+    killall udhcpc default.script dhcpcd 2>/dev/null
+    ifconfig "${INTERFACE}" 0.0.0.0
+fi
